MainPage tem seus valores preenchidos programaticamente

+++ Loja/Produto.cs
@@ -9,11 +9,11 @@ namespace Loja
     class Produto
     {
         public string Nome { get; set; }
-        public double Valor { get; set; }
+        public decimal Valor { get; set; }
         public string Descricao { get; set; }
         public string ImgSource { get; set; }
 
-        public Produto(string nome, double valor, string descricao, string imgSource)
+        public Produto(string nome, decimal valor, string descricao, string imgSource)
         {
             this.Nome = nome;
             this.Valor = valor;
